scripts/create_data/requirements.txt
anyio==4.7.0
astroid==3.3.6
black==23.7.0
blinker==1.9.0
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dill==0.3.9
exceptiongroup==1.2.2
Faker==33.1.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kafka-python==2.0.2
logger==1.4
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
py4j==0.10.9.7
pyactiveresource==2.2.2
PyJWT==2.10.1
pylint==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
schedule==1.2.2
ShopifyAPI==12.7.0
six==1.17.0
sniffio==1.3.1
stripe==11.3.0
tabulate==0.9.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.21.0